Crate holochain_net

holochain_net is a library that defines an abstract networking layer for different network transports, providing a configurable interface for swapping different backends connection modules at load time

Re-exports

pub extern crate holochain_core_types_derive;

Modules

connection
Provides a lightweight concurrency abstraction for holochain networking / p2p layer see holochain_net::ipc for a specific implementation, and holochain_net for the crate that pulls the implementations together

error

in_memory
Module for the in-memory network module

ipc
IPC Abstraction for P2P networking

ipc_net_worker
provides a NetWorker implementation for backend IPC p2p connections

p2p_config

p2p_network
This module provides the main abstraction for differing p2p backends P2pNetwork instances take a json configuration string and at load-time instantiate the configured "backend"

tweetlog

Macros

bail_generic
Macro akin to bail!() but returns an IpcError::GenericError.

log_d

log_dd

log_e

log_ee

log_i

log_ii

log_t

log_tt

log_w

log_ww