Enum holochain_net::connection::json_protocol::JsonProtocol

pub enum JsonProtocol {
    SuccessResult(SuccessResultData),
    FailureResult(FailureResultData),
    TrackDna(TrackDnaData),
    UntrackDna(TrackDnaData),
    Connect(ConnectData),
    PeerConnected(PeerData),
    GetState,
    GetStateResult(StateData),
    GetDefaultConfig,
    GetDefaultConfigResult(ConfigData),
    SetConfig(ConfigData),
    SendMessage(MessageData),
    SendMessageResult(MessageData),
    HandleSendMessage(MessageData),
    HandleSendMessageResult(MessageData),
    FetchEntry(FetchEntryData),
    FetchEntryResult(FetchEntryResultData),
    HandleFetchEntry(FetchEntryData),
    HandleFetchEntryResult(FetchEntryResultData),
    PublishEntry(EntryData),
    HandleStoreEntry(EntryData),
    HandleDropEntry(DropEntryData),
    FetchMeta(FetchMetaData),
    FetchMetaResult(FetchMetaResultData),
    HandleFetchMeta(FetchMetaData),
    HandleFetchMetaResult(FetchMetaResultData),
    PublishMeta(DhtMetaData),
    HandleStoreMeta(DhtMetaData),
    HandleDropMeta(DropMetaData),
    HandleGetPublishingEntryList(GetListData),
    HandleGetPublishingEntryListResult(EntryListData),
    HandleGetHoldingEntryList(GetListData),
    HandleGetHoldingEntryListResult(EntryListData),
    HandleGetPublishingMetaList(GetListData),
    HandleGetPublishingMetaListResult(MetaListData),
    HandleGetHoldingMetaList(GetListData),
    HandleGetHoldingMetaListResult(MetaListData),
}

Enum holding all message types that serialize as json in the 'hc-core <-> P2P network module' protocol. There are 4 categories of messages:

Command: An order from the local node to the p2p module. Local node expects a reponse. Starts with a verb.

Handle-command: An order from the p2p module to the local node. The p2p module expects a response. Start withs 'Handle' followed by a verb.

Result: A response to a Command. Starts with the name of the Command it responds to and ends with 'Result'.

Notification: Notify that something happened. Not expecting any response. Ends with verb in past form, i.e. '-ed'. Fetch = Request between node and the network (other nodes) Get = Request within a node between p2p module and core

Variants
SuccessResult(SuccessResultData)

Success response to a request (any message with an _id field.)

FailureResult(FailureResultData)

Failure response to a request (any message with an _id field.) Can also be a response to a mal-formed request.

TrackDna(TrackDnaData)

Order the p2p module to be part of the network of the specified DNA.

UntrackDna(TrackDnaData)

Order the p2p module to leave the network of the specified DNA.

Connect(ConnectData)

Connect to the specified multiaddr

PeerConnected(PeerData)

Notification of a connection from another peer.

GetState

Request the current state from the p2p module

GetStateResult(StateData)

p2p module's current state response.

GetDefaultConfig

Request the default config from the p2p module

GetDefaultConfigResult(ConfigData)

the p2p module's default config response

SetConfig(ConfigData)

Set the p2p config

SendMessage(MessageData)

Send a message to another peer on the network

SendMessageResult(MessageData)

the response from a previous SendMessage

HandleSendMessage(MessageData)

Request to handle a message another peer has sent us.

HandleSendMessageResult(MessageData)

Our response to a message from another peer.

FetchEntry(FetchEntryData)

Request data from the dht network

FetchEntryResult(FetchEntryResultData)

Response from requesting dht data from the network

HandleFetchEntry(FetchEntryData)

Another node, or the network module itself is requesting data from us

HandleFetchEntryResult(FetchEntryResultData)

Successful data response for a HandleFetchDhtData request

PublishEntry(EntryData)

Publish data to the dht.

HandleStoreEntry(EntryData)

Store data on a node's dht slice.

HandleDropEntry(DropEntryData)FetchMeta(FetchMetaData)

Request metadata from the dht

FetchMetaResult(FetchMetaResultData)

Response by the network for our metadata request

HandleFetchMeta(FetchMetaData)

Another node, or the network module itself, is requesting data from us

HandleFetchMetaResult(FetchMetaResultData)

Successful metadata response for a HandleFetchMeta request

PublishMeta(DhtMetaData)

Publish metadata to the dht.

HandleStoreMeta(DhtMetaData)

Store metadata on a node's dht slice.

HandleDropMeta(DropMetaData)

Drop metadata on a node's dht slice.

HandleGetPublishingEntryList(GetListData)HandleGetPublishingEntryListResult(EntryListData)HandleGetHoldingEntryList(GetListData)HandleGetHoldingEntryListResult(EntryListData)HandleGetPublishingMetaList(GetListData)HandleGetPublishingMetaListResult(MetaListData)HandleGetHoldingMetaList(GetListData)HandleGetHoldingMetaListResult(MetaListData)