Trait holochain_net::connection::net_connection::NetWorker

pub trait NetWorker {
    fn receive(&mut self, _data: Protocol) -> NetResult<()> { ... }

    fn tick(&mut self) -> NetResult<bool> { ... }

    fn stop(self: Box<Self>) -> NetResult<()> { ... }

    fn endpoint(&self) -> Option<String> { ... }
}

Trait that represents a worker thread that relays incoming and outgoing protocol messages between a handler closure and a p2p module.

Provided methods

fn receive(&mut self, _data: Protocol) -> NetResult<()>

The receiving method when NetSend's send() is called. It should relay that Protocol message to the p2p module.

fn tick(&mut self) -> NetResult<bool>

Polls the p2p module for Protocol messages received from the network, and perform any other upkeep. It should realy those messages back to the handler closure. Return false if no particular upkeep has been processed

fn stop(self: Box<Self>) -> NetResult<()>

stop the worker

fn endpoint(&self) -> Option<String>

Getter of the connection's endpoint

Loading content...